+++ BlinkyTest/src/Modules/WifiModule.h
@@ -2,13 +2,13 @@
 #include <Arduino.h>
 #include <vector>
 #include <SoftwareSerial.h>
-#include "Modules/SDCardModule.h"
+#include "./SDCardModule.h"
 
 class WifiModule: public SoftwareSerial
 {
   public:
      WifiModule(SDCardModule* sd, uint8_t rxPin, uint8_t txPin);
-     void process(uint8_t count);
+     void process(uint64_t count);
      void begin();
 
   private:
